# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión de Proyectos",
        "version": "1.0"
    },
    "paths": {
        "/api/projects": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Lista todos los proyectos",
                "operationId": "8d4a13ef3fc773ff4031e515460919ea",
                "responses": {
                    "200": {
                        "description": "Listado de proyectos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Crea un nuevo proyecto",
                "operationId": "d0613a552f223187f745c8fb72a8c189",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "team_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proyecto creado"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Obtiene un proyecto específico",
                "operationId": "af32aed47c19d517816e0b970d60cd5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proyecto encontrado"
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Actualiza un proyecto existente",
                "operationId": "1d36dcb3fa77a8da9800d82f0cfb70af",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "team_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proyecto actualizado"
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Elimina un proyecto",
                "operationId": "187865e30ff6b72f1cea3037d9dd84fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Proyecto eliminado"
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                }
            }
        },
...
    },
...
}